requirements.txt
aiodns==3.5.0
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==2.0.1
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
Deprecated==1.2.18
distro==1.9.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
netaddr==1.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycares==4.11.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.43
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
yookassa==3.7.1
